Former-commit-id: 0cce6a53197c3698fb81b644b40846fcc2bf8170

+++ Sources/Internal/Collision/Collisions.h
@@ -81,7 +81,7 @@ public:
         {
             avC += contactPoints[c];
         }
-        avC /= (float32)count;
+        avC /= count;
         return avC;
     };
 };
